backend/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv

# AI/ML dependencies
groq>=0.13.0
xgboost
lightgbm
imbalanced-learn
scikit-learn
shap
joblib
pandas
numpy

# OCR and image processing
rapidocr-onnxruntime
opencv-python
Pillow
PyMuPDF
rapidfuzz
python-multipart

# Blockchain and document generation
reportlab
qrcode[pil]
cryptography

# Additional utilities
pydantic>=2.0.0
httpx==0.27.0
